+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 * Fixed an issue where downloading indicator layoutsconatiner objects failed.
 * Added a validation that makes sure outputs in integrations are matching the README file.
-* Fixed an issue when files were classified incorrectly when running *update-release-notes*.
+* Fixed an issue when files were classified incorrectly when running **update-release-notes**.
 * Fixed an issue where the *create-id-set* flag in the **validate** command did not work.
 * Added a warning in case no id_set file is found when running the **validate** command.
+* Fixed an issue where changed files were not recognised correctly on forked branches in the **validate** and the **update-release-notes** commands.
 
 # 1.2.19
 * Fixed an issue where merge id_set was not updated to work with the new entity of Packs.